refactor(maencof): reorganize exports and update imports for clarity
Consolidate exports in the MCP modules by moving tool handlers to a centralized index file. Update import paths for graph-cache and related modules to reflect the new directory structure. Remove the edge-builders.ts file and distribute its functionality across new builder files for better organization.

- Updated exports in index.ts to streamline access to tools.
- Refactored imports in various test files to align with the new structure.
- Removed edge-builders.ts and created separate files for cross-layer and domain edge building logic.

This restructuring enhances code maintainability and clarity without altering existing functionality.
