+++ packages/maencof/src/core/graph-builder/index.ts
@@ -1,2 +1,1 @@
 export * from './graph-builder.js';
-export * from './edge-builders.js';
