+++ packages/maencof/src/mcp/server/graph-cache/graph-cache.ts
@@ -7,11 +7,11 @@
 import { homedir } from 'node:os';
 import { resolve } from 'node:path';
 
-import { MetadataStore } from '../../core/indexer/index.js';
-import { invalidateQueryCache } from '../../search/query-engine/index.js';
-import type { KnowledgeGraph } from '../../types/graph.js';
+import { MetadataStore } from '../../../core/indexer/index.js';
+import { invalidateQueryCache } from '../../../search/query-engine/index.js';
+import type { KnowledgeGraph } from '../../../types/graph.js';
 
-import { ensureFreshGraphNonBlocking } from './middlewares/freshness-guard.js';
+import { ensureFreshGraphNonBlocking } from '../middlewares/freshness-guard.js';
 
 /** Blocked prefixes for global config path access */
 const BLOCKED_PREFIXES = [
