+++ packages/maencof/src/mcp/server/graph-cache/index.ts
@@ -1,0 +1,6 @@
+export {
+  getVaultPath,
+  loadGraphIfNeeded,
+  ensureFreshGraph,
+  invalidateCache,
+} from './graph-cache.js';
