+++ packages/maencof/src/mcp/server/middlewares/background-rebuild.ts
@@ -6,7 +6,7 @@
 import { appendErrorLogSafe } from '../../../core/error-log/index.js';
 import { handleKgBuild } from '../../tools/kg-build/index.js';
 
-import { invalidateCache } from '../graph-cache.js';
+import { invalidateCache } from '../graph-cache/index.js';
 
 // 단일 vault 가정. multi-vault 도입 시 Map<vaultPath, Promise<void>>로 전환.
 let rebuildInProgress: Promise<void> | null = null;
